requirements.txt
fastapi
uvicorn
pydantic
requests
streamlit
python-dotenv
openai
google-generativeai
